requirements.txt
diffusers==0.30.2
matplotlib==3.9.2
Pillow==10.4.0
streamlit==1.38.0
torch==2.4.1
